% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot2.R
\name{boxplot2_binding}
\alias{boxplot2_binding}
\title{Basic boxplot plotter. code can be modified before plotting with format edit, and bound in the global}
\usage{
boxplot2_binding(x, format = "plot")
}
\arguments{
\item{x}{data in the form of a dbl vector for which you would like the make a boxplot}

\item{format}{Whether the function should plot, or return an object for editing.}

\item{name}{what name will the list object containing your plot code and environment contain}
}
\value{
a ggplot object or a named list containing ggplot code along with a function environment.
}
\description{
Basic boxplot plotter. code can be modified before plotting with format edit, and bound in the global
}
\examples{
a <- rbeta(1e2,1,3)
boxplot2_binding(x = a) # just makes a boxplot
diamond_points <- boxplot2_binding(x = a, format = "edit") # opens a doc for editing
diamaond_points #runs the edited code using new S3 method

}
