% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot2.R
\name{boxplot2_name}
\alias{boxplot2_name}
\title{Basic boxplot plotter. code can be modified before plotting with format edit}
\usage{
boxplot2_name(x, format = "plot", name = "code_output")
}
\arguments{
\item{x}{data in the form of a dbl vecotor for which you would like the make a boxplot}

\item{format}{Whether the function should plot, or return an object for editing.}

\item{name}{what name will the list object containing your plot code and environment contain}
}
\value{
a ggplot object or a named list containing ggplot code along with a function environment.
}
\description{
Basic boxplot plotter. code can be modified before plotting with format edit
}
\examples{
a <- rbeta(1e2,1,3)
boxplot2_name(x = a) # just makes a boxplot
boxplot2_name(x = a, formate = "edit", name = "circle_points") # opens a doc for editing
boxplot2_adjust("circle_points") #runs the edited code

}
