% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot2.R
\name{boxplot2_original}
\alias{boxplot2_original}
\title{Basic boxplot plotter.  code can be edited before plotting with format edit. boxplot2_adjust can run the modified code.}
\usage{
boxplot2_original(x, format = "plot")
}
\arguments{
\item{x}{data in the form of a dbl vecotor for which you would like the make a boxplot}

\item{format}{Whether the function should plot, or return an object for editing.}
}
\value{
a ggplot object or a list containing ggplot code along with a function environment.
}
\description{
Basic boxplot plotter.  code can be edited before plotting with format edit. boxplot2_adjust can run the modified code.
}
\examples{
a <- rbeta(1e2,1,3)
boxplot2_original(x = a) # just makes a boxplot
boxplot2_original(x = a, formate = "edit") # opens a doc for editing
boxplot2_adjust() #runs the edited code

}
