% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot2.R
\name{boxplot2_template}
\alias{boxplot2_template}
\title{Basic boxplot plotter.  Templates can be made to adjust it with make_template.}
\usage{
boxplot2_template(x, template = NULL)
}
\arguments{
\item{x}{data in the form of a dbl vecotor for which you would like the make a boxplot}

\item{template}{if you have made a template with make_template, placing its name as a string here will run the template code instead.}
}
\value{
a ggplot object
}
\description{
Basic boxplot plotter.  Templates can be made to adjust it with make_template.
}
\examples{
make_template("boxplot2", "triangle") # make any desired changes to the boxplot
a <- rbeta(1e2,1,3)
boxplot2_template(x = a, template = "triangle.R")

}
