% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_binding}
\alias{boxplot_binding}
\title{Basic boxplot plotter. code can be modified before plotting with format edit, and bound in the global}
\usage{
boxplot_binding(x, format = "plot")
}
\arguments{
\item{x}{data in the form of a tibble with a column of values named value and column of classes name class}

\item{format}{Whether the function should plot, or return an object for editing.}
}
\value{
a ggplot object or a bnd class list containing ggplot code along with a function environment.
}
\description{
Basic boxplot plotter. code can be modified before plotting with format edit, and bound in the global
}
\examples{
a <- tibble(value = c(rnorm(2e2,1), 4*rbeta(2e2,1,4)), class = c(rep("a",2e2),rep("b",2e2)))
boxplot_binding(x = a) # just makes a boxplot
diamond_points <- boxplot_binding(x = a, format = "edit") # opens a doc for editing
diamond_points #runs the edited code using new S3 method

}
