% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_name}
\alias{boxplot_name}
\title{Basic boxplot plotter. code can be modified before plotting with format edit}
\usage{
boxplot_name(x, format = "plot", name = "code_output")
}
\arguments{
\item{x}{data in the form of a tibble with a column of values named value and column of classes name class}

\item{format}{Whether the function should plot (plot), or return an object for editing (edit).}

\item{name}{what name will the list object containing your plot code and environment contain}
}
\value{
a ggplot object or a named list containing ggplot code along with a function environment.
}
\description{
Basic boxplot plotter. code can be modified before plotting with format edit
}
\examples{
a <- tibble(value = c(rnorm(2e2,1), 4*rbeta(2e2,1,4)), class = c(rep("a",2e2),rep("b",2e2)))
boxplot_name(x = a) # just makes a boxplot
boxplot_name(x = a, format = "edit", name = "circle_points") # opens a doc for editing
boxplot_adjust(circle_points) #runs the edited code

}
