% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_original}
\alias{boxplot_original}
\title{Basic boxplot plotter.  code can be edited before plotting with format edit. boxplot_adjust can run the modified code.}
\usage{
boxplot_original(x, format = "plot")
}
\arguments{
\item{x}{data in the form of a dbl vector for which you would like the make a boxplot}

\item{format}{Whether the function should plot, or return an object for editing.}
}
\value{
a ggplot object or a list containing ggplot code along with a function environment.
}
\description{
Basic boxplot plotter.  code can be edited before plotting with format edit. boxplot_adjust can run the modified code.
}
\examples{
a <- tibble(value = c(rnorm(2e2,1), 4*rbeta(2e2,1,4)), class = c(rep("a",2e2),rep("b",2e2)))
boxplot_original(x = a) # just makes a boxplot
boxplot_original(x = a, format = "edit") # opens a doc for editing
boxplot_adjust() #runs the edited code

}
