\docType{data}
\name{Cameca}
\alias{Cameca}
\title{Example datasets for testing \code{simplex}}
\description{
  a Plesovice and Qinghu zircon dataset from the Cameca SIMS instrument
  at the Chinese Academy of Sciences Beijing.
}
\usage{Cameca}
\format{An object of class \code{simplex}.}
\keyword{datasets}