% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Cameca_oxygen}
\alias{Cameca_oxygen}
\title{example oxygen isotope dataset}
\description{
built-in Cameca oxigen dataset containing NBS28,
    Qinghu zircon and ZBL10 data
}
\examples{
data(Cameca_oxygen,package="simplex")
plot(Cameca_oxygen,i=1)
}
