% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Cameca_UPb}
\alias{Cameca_UPb}
\title{example Cameca dataset}
\description{
built-in Cameca U-Pb dataset containing Plesovice and
    Qinghu zircon
}
\examples{
data(Cameca_UPb,package="simplex")
plot(Cameca_UPb,i=1)
}
