% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{data2table}
\alias{data2table}
\alias{data2table.default}
\alias{data2table.calibrated}
\alias{data2table.delta}
\title{export to table}
\usage{
data2table(x, ...)

\method{data2table}{default}(x, ...)

\method{data2table}{calibrated}(x, ...)

\method{data2table}{delta}(x, ...)
}
\arguments{
\item{x}{an object of class \code{calibrated} or \code{delta}}

\item{...}{optional arguments}
}
\value{
a matrix
}
\description{
convert calibrated SIMS data objects to a flat table
}
\examples{
data('Cameca',package='simplex')
dc <- drift(x=Cameca)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=standard(preset='Plesovice'))
cd <- calibrate(cal)
tab <- data2table(cd)
}
