% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{delta}
\alias{delta}
\title{calculate delta values}
\usage{
delta(cd, ref, log = TRUE)
}
\arguments{
\item{cd}{an object of class \code{calibrated}}

\item{ref}{the logratio composition of a reference material, i.e. a
list with items \code{val} (vector of logratios) and \code{cov}
(the covariance matrix of \code{val})}

\item{log}{use traditional delta notation (\code{log=FALSE}) or the
logratio definition (\code{log=TRUE})?}
}
\value{
an object of class \code{delta}
}
\description{
convert stable isotope logratio data to delta notation
}
\examples{
\dontrun{
data('Cameca_oxygen',package='simplex')
dc <- drift(x=Cameca_oxygen)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=standard(preset='NBS28-O'))
cd <- calibrate(cal)
del <- delta(cd)
tab <- data2table(del)
}
}
