% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratios.R
\name{getPbLogRatios}
\alias{getPbLogRatios}
\title{estimate atomic Pb-Pb logratios from count data}
\usage{
getPbLogRatios(dat)
}
\arguments{
\item{dat}{an object of class \code{simplex}}
}
\value{
a list with the following items:

\code{x}: a vector of logratios

\code{cov}: the covariance matrix of \code{x}

\code{snames}: the sample names

\code{num}: the numerator isotopes of \code{x}

\code{den}: the denominator isotopes of \code{x}
}
\description{
turns Pb count data into atomic logratios, using a
    binomial likelihood model
}
\examples{
data(Cameca,package="simplex")
stand <- standards(dat=Cameca,prefix='Plesovice',tst=c(337.13,0.18))
calU <- calibration(stand=stand,oxide='UO2',parent='U238',
                    daughter='Pb206',cD4=18.7)
samp <- unknowns(dat=Cameca,prefix='Qinghu')
calUsamp <- calibrate(dat=samp,fit=calU)
calPbsamp <- getPbLogRatios(samp)
calsamp <- mergecal(calUsamp,calPbsamp)
tab <- data2table(calsamp)
}
