% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratios.R
\name{logratios}
\alias{logratios}
\title{logratio calculation}
\usage{
logratios(x)
}
\arguments{
\item{x}{an object of class \code{drift}}
}
\value{
an object of class \code{logratios}
}
\description{
extracts logratio averages from time resolved SIMS data
}
\examples{
\dontrun{
data('SHRIMP',package='simplex')
dc <- drift(x=SHRIMP)
lr <- logratios(dc)
plot(lr,i=1,option=2)
}
}
