% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{mergecal}
\alias{mergecal}
\title{merge calibrated data}
\usage{
mergecal(...)
}
\arguments{
\item{...}{any number of objects of class \code{calibrated}}
}
\value{
a new object of class \code{calibrated}
}
\description{
groups calibrated U-Pb, Th-Pb and Pb-Pb ratios into
    one object for further processing and plotting
}
\examples{
data(Cameca,package="simplex")
stand <- standards(dat=Cameca,prefix='Plesovice',tst=c(337.13,0.18))
calU <- calibration(stand=stand,oxide='UO2',parent='U238',
                    daughter='Pb206',cD4=18.7)
samp <- unknowns(dat=Cameca,prefix='Qinghu')
calUsamp <- calibrate(dat=samp,fit=calU)
calPbsamp <- getPbLogRatios(samp)
calsamp <- mergecal(calUsamp,calPbsamp)
tab <- data2table(calsamp)
}
