% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{multicollector}
\alias{multicollector}
\title{Test if a dataset was measured by multicollector mass spectrometry}
\usage{
multicollector(x, ...)
}
\arguments{
\item{x}{an simplex data object}

\item{...}{optional arguments (not used)}
}
\description{
Test if a dataset was measured by multicollector mass spectrometry
}
