% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{oxygen}
\alias{oxygen}
\title{example oxygen isotope dataset}
\description{
built-in Cameca oxigen dataset containing NBS28,
    Qinghu zircon and ZBL10 data
}
\examples{
data(oxygen,package="simplex")
plot(oxygen,i=1)
}
