% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{plot.calibrated}
\alias{plot.calibrated}
\title{plot calibrated data}
\usage{
\method{plot}{calibrated}(x, type, ...)
}
\arguments{
\item{x}{an object of class \code{calibrated}}

\item{type}{for U-Pb or U-Th-Pb data, either \code{'U-Pb'} or
\code{'Th-Pb'}.}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function}
}
\description{
shows the calibrated data on a logratio plot.
}
\examples{
data('Cameca')
dc <- drift(x=Cameca)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=standard(preset='Plesovice'))
cd <- calibrate(cal)
plot(cd,type='U-Pb')
}
