% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{plot.calibration}
\alias{plot.calibration}
\title{plot calibration data}
\usage{
\method{plot}{calibration}(x, show.numbers = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{calibration}}

\item{show.numbers}{logical. If \code{TRUE}, labels the error
ellipses with the aliquot numbers.}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function.}
}
\description{
shows the calibration data on a logratio plot.
}
\examples{
\dontrun{
data('SHRIMP_UPb',package='simplex')
st <- standard(preset='Temora')
dc <- drift(x=SHRIMP_UPb)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=st,prefix='TEM')
plot(cal)
}
}
