% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{plot.delta}
\alias{plot.delta}
\title{plot delta values}
\usage{
\method{plot}{delta}(x, ...)
}
\arguments{
\item{x}{an object of class \code{delta}}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function.}
}
\description{
visualises stable isotope data as error ellipses in
    delta notation
}
\examples{
data('Cameca_oxygen',package='simplex')
dc <- drift(x=Cameca_oxygen)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=standard(preset='NBS28'))
cd <- calibrate(cal)
del <- delta(cd)
plot(del)
}
