% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{plot.drift}
\alias{plot.drift}
\title{plot drift corrected data}
\usage{
\method{plot}{drift}(x, sname = NULL, i = 1, ...)
}
\arguments{
\item{x}{an object of class \code{drift}}

\item{sname}{the name of the sample to be shown}

\item{i}{the number of the sample to be shown}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function.}
}
\description{
shows the time resolved mass spectrometer signals
    fitted by a generalised linear model
}
\examples{
data('Cameca_UPb',package='simplex')
dc <- drift(x=Cameca_UPb)
plot(dc,i=1)
}
