% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plot.simplex}
\alias{plot.simplex}
\title{plot simplex data}
\usage{
\method{plot}{simplex}(x, sname = NULL, i = 1, ...)
}
\arguments{
\item{x}{an object of class \code{simplex}}

\item{sname}{the sample name to be shown}

\item{i}{the sample number to be shown}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function.}
}
\description{
plot time resolved SIMS data
}
\examples{
data('SHRIMP_UPb',package='simplex')
plot(SHRIMP_UPb,i=1)
}
