% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeresolved}
\alias{plot_timeresolved}
\title{plot the raw mass spectrometer data}
\usage{
plot_timeresolved(samp, fit = NULL)
}
\arguments{
\item{samp}{one item of a \code{simplex} list object}

\item{fit}{the output of \code{calibration}}
}
\value{
a multi-panel plot
}
\description{
Shows the raw data for a single spot in a SIMS dataset.
}
\examples{
data(Cameca,package="simplex")
plot_timeresolved(Cameca[[1]])
}
