% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process SIMS data}
\usage{
process(f, method, stand, t = 0, exterr = FALSE)
}
\arguments{
\item{f}{file name(s), may include wildcards (\code{*.asc},
\code{*.op} or \code{*.pd}).}

\item{method}{an object of class \code{method} OR the name of a
data acquisition protocol (one of \code{'IGG-UPb'},
\code{'GA-UPb'}, \code{'IGG-UThPb'}, \code{'IGG-O'}, or
\code{'IGG-S'}). To create new methods, see \code{method}.}

\item{stand}{an object of class \code{standard}}

\item{t}{the analysis time to which the logratio signals should be
regressed.}

\item{exterr}{include the uncertainty associated with the
standard calibration in the error propagation?}
}
\value{
an object of class \code{calibrated}
}
\description{
one stop shop for SIMS data reduction, including
    including the drift correction, logratio calculation, and
    calibration.
}
\examples{
\dontrun{
m <- method('GA-UPb')
s <- standard(preset="Temora",prefix='TEM')
fname <- system.file('SHRIMP.pd',package='simplex')
cd <- process(f=fname,method=m,stand=s)
tab <- data2table(cd)
}
}
