% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud_binding}
\alias{raincloud_binding}
\title{Raincloud-style distribution plotter. code can be modified before plotting with format edit, and bound in the global.}
\usage{
raincloud_binding(x, format = "plot")
}
\arguments{
\item{x}{data in the form of a tibble with a column of values named value and column of classes named class.}

\item{format}{Whether the function should plot, or return an object for editing.}
}
\value{
a ggplot object or a bnd class list containing ggplot code along with a function environment.
}
\description{
Raincloud-style distribution plotter. code can be modified before plotting with format edit, and bound in the global.
}
\examples{
a <- tibble(value = c(rnorm(2e2, 1), 4 * rbeta(2e2, 1, 4)),
            class = c(rep("a", 2e2), rep("b", 2e2)))
raincloud_binding(x = a)
custom_points <- raincloud_binding(x = a, format = "edit")
custom_points

}
