% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud_name}
\alias{raincloud_name}
\title{Raincloud-style distribution plotter. code can be modified before plotting with format edit.}
\usage{
raincloud_name(x, format = "plot", name = "code_output")
}
\arguments{
\item{x}{data in the form of a tibble with a column of values named value and column of classes named class.}

\item{format}{Whether the function should plot (plot), or return an object for editing (edit).}

\item{name}{what name will the list object containing your plot code and environment contain.}
}
\value{
a ggplot object or a named list containing ggplot code along with a function environment.
}
\description{
Raincloud-style distribution plotter. code can be modified before plotting with format edit.
}
\examples{
a <- tibble(value = c(rnorm(2e2, 1), 4 * rbeta(2e2, 1, 4)),
            class = c(rep("a", 2e2), rep("b", 2e2)))
raincloud_name(x = a)
raincloud_name(x = a, format = "edit", name = "transparent_points")
raincloud_adjust(transparent_points)

}
