% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud_original}
\alias{raincloud_original}
\title{Raincloud-style distribution plotter. code can be edited before plotting with format edit. raincloud_adjust can run the modified code.}
\usage{
raincloud_original(x, format = "plot")
}
\arguments{
\item{x}{data in the form of a tibble with a column of values named value and column of classes named class.}

\item{format}{Whether the function should plot, or return an object for editing.}
}
\value{
a ggplot object or a list containing ggplot code along with a function environment.
}
\description{
Raincloud-style distribution plotter. code can be edited before plotting with format edit. raincloud_adjust can run the modified code.
}
\examples{
a <- tibble(value = c(rnorm(2e2, 1), 4 * rbeta(2e2, 1, 4)),
            class = c(rep("a", 2e2), rep("b", 2e2)))
raincloud_original(x = a)
raincloud_original(x = a, format = "edit")
raincloud_adjust()

}
