% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud_template}
\alias{raincloud_template}
\title{Raincloud-style distribution plotter. Templates can be made to adjust it with make_template.}
\usage{
raincloud_template(x, template = NULL)
}
\arguments{
\item{x}{data in the form of a tibble with a column of values named value and column of classes named class.}

\item{template}{if you have made a template with make_template, placing its name as a string here will run the template code instead.}
}
\value{
a ggplot object
}
\description{
Raincloud-style distribution plotter. Templates can be made to adjust it with make_template.
}
\examples{
make_template("raincloud", "custom_raincloud")
a <- tibble(value = c(rnorm(2e2, 1), 4 * rbeta(2e2, 1, 4)),
            class = c(rep("a", 2e2), rep("b", 2e2)))
raincloud_template(x = a, template = "custom_raincloud.R")

}
