% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_data}
\alias{read_data}
\title{read SIMS data}
\usage{
read_data(dorf, instrument = "Cameca", suffix = NULL)
}
\arguments{
\item{dorf}{directory or file name}

\item{instrument}{either \code{'Cameca'} or \code{'SHRIMP'}}

\item{suffix}{the extension of the data files to be loaded. This
defaults to \code{.asc} if \code{instrument='Cameca'} and
\code{.op} if \code{instrument='SHRIMP'}.}
}
\value{
an object of class \code{simplex}
}
\description{
read a file or folder with ASCII data
}
\examples{
# not run:
\dontrun{
camdat <- read_data('/path/to/asc/files/',instrument='Cameca',suffix='.asc')
plot_timeresolved(camdat[[1]])
}
}
