% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{SHRIMP}
\alias{SHRIMP}
\title{example SHRIMP dataset}
\description{
built-in SHRIMP U-Pb dataset containing Temora and
    91500 zircon
}
\examples{
data(SHRIMP,package="simplex")
plot(SHRIMP,i=1)
}
