% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{SHRIMP_UPb}
\alias{SHRIMP_UPb}
\title{example SHRIMP dataset}
\description{
built-in SHRIMP U-Pb dataset containing Temora and
    91500 zircon
}
\examples{
data(SHRIMP_UPb,package="simplex")
plot(SHRIMP_UPb,i=1)
}
