\name{simplex-package}
\alias{simplex-package}
\alias{simplex}
\docType{package}
\title{
\packageTitle{simplex}
}
\description{
\packageDescription{simplex}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simplex}
\packageIndices{simplex}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{simplex}

Maintainer: \packageMaintainer{simplex}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
