% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{simplex}
\alias{simplex}
\alias{_PACKAGE}
\alias{simplex-package}
\title{simplex: a SIMS data reduction package}
\description{
\code{simplex} is a data reduction package for Secondary Ion Mass
Spectrometry (SIMS), with functionality for both stable isotope
geochemistry and U-(Th)-Pb geochronology. The package accommodates
data from both Cameca and SHRIMP instruments. All mathematical
operations take place in logratio space (aka the simplex) to ensure
positivity and keep track of correlated uncertainties.
}
\details{
A list of documented functions may be viewed by typing
\code{help(package='simplex')}.  Source code and further
instructions are provided at
\url{https://github.com/pvermees/simplex/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pvermees/simplex}
}

}
\author{
\strong{Maintainer}: Pieter Vermeesch \email{p.vermeesch@ucl.ac.uk}

}
