% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{simplex2IsoplotR}
\alias{simplex2IsoplotR}
\title{convert to IsoplotR}
\usage{
simplex2IsoplotR(dat, method = "U-Pb")
}
\arguments{
\item{dat}{an object of class \code{calibrated}}

\item{method}{sets the format of the \code{IsoplotR} object. If
\code{'U-Pb'}, produces a format 5 object of class \code{UPb};
if \code{'U-Th-Pb'}, produces a format 8 object of class
\code{UPb}; if \code{'Th-Pb'}, produces a format 2 object of
class \code{ThPb}; if \code{'Pb-Pb'}, produces a format 2
object of class \code{PbPb}.}
}
\value{
an object of class \code{UPb}, \code{ThPb} or \code{PbPb}
}
\description{
convert U-Pb, U-Th-Pb, or Pb-Pb data to an IsoplotR
    object
}
\examples{
\dontrun{
m <- method('GA-UPb')
s <- standard(preset="Temora",prefix='TEM')
fname <- system.file('SHRIMP.pd',package='simplex')
cd <- process(f=fname,m=m,stand=s)
UPb <- simplex2IsoplotR(cd)
IsoplotR::concordia(UPb)
}
}
