% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{standards}
\alias{standards}
\title{define the standards in a dataset}
\usage{
standards(dat, prefix, invert = FALSE, c64 = 18.7, PbU = NULL,
  tst = NULL)
}
\arguments{
\item{dat}{an object of class \code{simplex}}

\item{prefix}{text string to match}

\item{invert}{logical.  If \code{TRUE} return samples whose names
do _not_ match}

\item{c64}{the \eqn{^{206}}Pb/\eqn{^{204}}Pb-ratio of the common Pb}

\item{PbU}{(optional) true \eqn{^{206}}Pb/\eqn{^{238}}U-ratio of
the age standard}

\item{tst}{(optional) two-element vector with the age and standard
error of the age standard}
}
\value{
an object of class \code{standard}
}
\description{
select a subset of standards from a \code{simplex}
    dataset.
}
\examples{
data(Cameca,package="simplex")
stand <- standards(dat=Cameca,prefix='Plesovice')
}
