% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{test-data}
\alias{test-data}
\alias{Cameca_UPb}
\alias{Cameca_UThPb}
\alias{Cameca_oxygen}
\alias{Cameca_sulphur}
\alias{SHRIMP_UPb}
\title{test data}
\format{
An object of class \code{calibrated} (inherits from \code{calibration}, \code{logratios}, \code{drift}, \code{simplex}) of length 5.

An object of class \code{calibrated} (inherits from \code{calibration}, \code{logratios}, \code{drift}, \code{simplex}) of length 5.

An object of class \code{calibrated} (inherits from \code{calibration}, \code{logratios}, \code{drift}, \code{simplex}) of length 5.

An object of class \code{calibrated} (inherits from \code{calibration}, \code{logratios}, \code{drift}, \code{simplex}) of length 5.

An object of class \code{calibrated} (inherits from \code{calibration}, \code{logratios}, \code{drift}, \code{simplex}) of length 5.
}
\usage{
Cameca_UPb

Cameca_UThPb

Cameca_oxygen

Cameca_sulphur

SHRIMP_UPb
}
\description{
Various Cameca and SHRIMP datasets
}
\keyword{datasets}
