% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{troilite}
\alias{troilite}
\title{troilite composition}
\usage{
troilite()
}
\value{
a two-item list containing a vector \code{val} of
    logratio-values and its covariance matrix \code{cov}.
}
\description{
returns the sulphur isotope composition of meteoritic
    troilite
}
\examples{
data(Cameca_sulphur,package="simplex")
del <- list()
del$val <- c(0.83,1.61,3.25)
del$cov <- diag(c(0.03,0.08,0.03)^2)
names(del$val) <- rownames(del$cov) <-
colnames(del$cov) <- c('S33/S32','S34/S32','S36/S32')
st <- standard(del=del,ref=troilite())
cal <- calibration(lr=Cameca_sulphur,stand=st)
plot(cal)
}
