% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{unknown}
\alias{unknown}
\title{define the samples in a dataset}
\usage{
unknown(dat, prefix, invert = FALSE)
}
\arguments{
\item{dat}{an object of class \code{simplex}}

\item{prefix}{text string to match}

\item{invert}{logical.  If \code{TRUE} return samples whose names
do _not_ match}
}
\value{
a list of objects of class \code{unknown}
}
\description{
select a subset of samples of unknown age from a
    \code{simplex} dataset.
}
\examples{
data(Cameca,package="simplex")
unk <- unknown(Cameca,prefix='Plesovice',invert=TRUE)
}
