% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{VSMOW}
\alias{VSMOW}
\title{Vienna Standard Mean Ocean Water}
\usage{
VSMOW()
}
\value{
a two-item list containing a vector \code{val} of
    logratio-values and its covariance matrix \code{cov}.
}
\description{
returns the oxygen isotope composition of the
    eponymous stable isotope reference material
}
\examples{
data(Cameca_oxygen,package="simplex")
del <- list()
del$val <- 5.4
del$cov <- matrix(0.04,1,1)
names(del$val) <- colnames(del$cov) <- rownames(del$cov) <- 'O18/O16'
st <- standard(del=del,ref=VSMOW())
cal <- calibration(lr=Cameca_oxygen,stand=st)
plot(cal)
}
