% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{Basic weather plotter.}
\usage{
weather(station = "USW00013959", yr = 2025, title = NULL)
}
\arguments{
\item{station}{The station you would like to plot weather data for}

\item{year}{the year you would like to plot}
}
\value{
a ggplot object
}
\description{
Basic weather plotter.
}
