% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_template}
\alias{boxplot_template}
\title{Basic boxplot plotter.  Templates can be made to adjust it with make_template.}
\usage{
boxplot_template(x, value, class, template = NULL)
}
\arguments{
\item{x}{data in the form of a tibble with a column of value and column of classes}

\item{value}{the name of the column of values}

\item{class}{the name of the column of classes}

\item{template}{if you have made a template with make_template, placing its name as a string here will run the template code instead.}
}
\value{
a ggplot object
}
\description{
Basic boxplot plotter.  Templates can be made to adjust it with make_template.
}
\examples{
make_template("boxplot", "triangle") # make any desired changes to the boxplot
a <- rbeta(1e2,1,3)
boxplot_template(x = a, template = "triangle.R")

}
