% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratios.R
\name{calibrate}
\alias{calibrate}
\title{apply a calibration to sample data}
\usage{
calibrate(dat, fit)
}
\arguments{
\item{dat}{a dataset of class \code{simplex}}

\item{fit}{the output of \code{calibration}}
}
\value{
an \code{IsoplotR} object of class \code{UPb}
    (\code{format=5})
}
\description{
fit a straight line to log[Pb/U] vs. log[UOx/U] sample
    data
}
\details{
Regresses a line through
    X=\eqn{^{238}}U\eqn{^{16}}O\eqn{_x}/\eqn{^{238}}U and
    Y=(\eqn{^{206}}Pb-\eqn{^{204}}Pb[\eqn{^{206}}Pb/
    \eqn{^{204}}Pb]\eqn{_{\circ}})/\eqn{^{238}}U using the same
    slope as previously determined on an age standard
}
\examples{
data(Cameca,package="simplex")
stand <- subset_samples(dat=Cameca,prefix='Plesovice')
fit <- calibration(stand,tst=c(337.13,0.18))
samp <- subset_samples(dat=Cameca,prefix='Qinghu')
cal <- calibrate(samp,fit)
}
