% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{calibrate SIMS data}
\usage{
calibrate(cal, exterr = FALSE)
}
\arguments{
\item{cal}{an object of class \code{calibration}}

\item{exterr}{logical flag indicating whether the uncertainty of
the calibration should be propagated.}
}
\value{
an object of class \code{calibrated}
}
\description{
convert signal logratios to atomic logratios
}
\examples{
data('oxygen')
dc <- drift(x=oxygen)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=standard(preset='NBS28'),prefix='NBS28')
cd <- calibrate(cal)
del <- delta(cd,log=FALSE)
tab <- data2table(del)
}
