% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratios.R
\name{calibrate}
\alias{calibrate}
\title{apply a calibration to sample data}
\usage{
calibrate(dat, fit, syserr = FALSE)
}
\arguments{
\item{dat}{a dataset of class \code{simplex}}

\item{fit}{the output of \code{calibration}}

\item{syserr}{propagate the systematic error associated with the
standard (age uncertainty and calibration fit)?}
}
\value{
an \code{IsoplotR} object of class \code{UPb}
    (\code{format=5})
}
\description{
fit a straight line to log[Pb/U] vs. log[UOx/U] sample
    data
}
\details{
Regresses a line through
    X=\eqn{^{238}}U\eqn{^{16}}O\eqn{_x}/\eqn{^{238}}U and
    Y=(\eqn{^{206}}Pb-\eqn{^{204}}Pb[\eqn{^{206}}Pb/
    \eqn{^{204}}Pb]\eqn{_{\circ}})/\eqn{^{238}}U using the same
    slope as previously determined on an age standard
}
\examples{
data(Cameca,package="simplex")
stand <- standards(dat=Cameca,prefix='Plesovice',tst=c(337.13,0.18))
fit <- calibration(stand,oxide='UO2')
unk <- unknowns(dat=Cameca,prefix='Plesovice',invert=TRUE)
cal <- calibrate(unk,fit)
}
