% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{standard calibration}
\usage{
calibration(
  lr,
  stand,
  pairing = NULL,
  prefix = NULL,
  snames = NULL,
  i = NULL,
  invert = FALSE,
  t = NULL
)
}
\arguments{
\item{lr}{an object of class \code{logratios}}

\item{stand}{an object of class \code{standard}}

\item{pairing}{an object of class \code{pairing}}

\item{prefix}{(optional) prefix of the aliquots to be used as
calibration standards}

\item{snames}{(option) vector of sample names to be used as
calibration standards}

\item{i}{(optional) vector of indices of aliquots to be used as
calibration standards}

\item{invert}{if \code{TRUE}, inverts the selection made by
\code{snames} or \code{i}}

\item{t}{analysis time that the signal should be regressed to}
}
\value{
an object of class \code{calibration}
}
\description{
calibration of SIMS data using reference standards
}
\examples{
\dontrun{
data('SHRIMP_UPb',package='simplex')
st <- standard(preset='Temora')
dc <- drift(x=SHRIMP_UPb)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=st,prefix="TEM")
plot(cal)
}
}
