% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{standard calibration}
\usage{
calibration(lr, stand, snames = NULL, i = NULL, invert = FALSE, t = NULL)
}
\arguments{
\item{lr}{an object of class \code{logratios}}

\item{stand}{an object of class \code{standard}}

\item{snames}{a vector with the sample names of selected standard
analyses to be used for the calibration}

\item{i}{a vector of indices of selected standard analyses to be
used for the calibration}

\item{invert}{if \code{TRUE}, inverts the selection made by
\code{snames} or \code{i}}

\item{t}{analysis time that the signal should be regressed to}
}
\value{
an object of class \code{calibration}
}
\description{
calibration of SIMS data using reference standards
}
\examples{
\dontrun{
data('SHRIMP',package='simplex')
st <- standard(preset='Temora',prefix=TEM)
dc <- drift(x=SHRIMP)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=st)
plot(cal,option=3)
}
}
