% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calplot}
\alias{calplot}
\title{calibration plot}
\usage{
calplot(dat, fit)
}
\arguments{
\item{dat}{an object of class \code{simplex}}

\item{fit}{the output of \code{calibration}}
}
\value{
a plot of Y=(\eqn{^{206}}Pb-\eqn{^{204}}Pb[\eqn{^{206}}Pb/
             \eqn{^{204}}Pb]\eqn{_{\circ}})/\eqn{^{238}}U
    versus X=\eqn{^{238}}U\eqn{^{16}}O\eqn{_x}/\eqn{^{238}}U
}
\description{
plot Y=(\eqn{^{206}}Pb-\eqn{^{204}}Pb[\eqn{^{206}}Pb/
        \eqn{^{204}}Pb]\eqn{_{\circ}})/\eqn{^{238}}U
    versus X=\eqn{^{238}}U\eqn{^{16}}O\eqn{_x}/\eqn{^{238}}U
}
\examples{
data(Cameca,package="simplex")
Ples <- standards(dat=Cameca,prefix='Plesovice')
cal <- calibration(stand=Ples,oxide='UO2')
calplot(dat=Ples,fit=cal)
}
