% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calplot}
\alias{calplot}
\title{calibration plot}
\usage{
calplot(stand, fit)
}
\arguments{
\item{stand}{an object of class \code{standard}}

\item{fit}{the output of \code{calibration}}
}
\value{
a plot of Y=(\eqn{^{206}}Pb-\eqn{^{204}}Pb[\eqn{^{206}}Pb/
             \eqn{^{204}}Pb]\eqn{_{\circ}})/\eqn{^{238}}U
    versus X=\eqn{^{238}}U\eqn{^{16}}O\eqn{_x}/\eqn{^{238}}U
}
\description{
plot Y=(\eqn{^{206}}Pb-\eqn{^{204}}Pb[\eqn{^{206}}Pb/
        \eqn{^{204}}Pb]\eqn{_{\circ}})/\eqn{^{238}}U
    versus X=\eqn{^{238}}U\eqn{^{16}}O\eqn{_x}/\eqn{^{238}}U
}
\examples{
data(Cameca,package="simplex")
Ples <- standard(dat=Cameca,prefix='Plesovice',tst=c(337.13,0.18))
cal <- calibration(stand=Ples,oxide='UO2')
calplot(stand=Ples,fit=cal)
}
