% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Cameca}
\alias{Cameca}
\title{example Cameca dataset}
\description{
built-in Cameca dataset containing Plesovice and
    Qinghu zircon
}
\examples{
data(Cameca,package="simplex")
plot_timeresolved(Cameca[[1]])
}
