% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{data2table}
\alias{data2table}
\title{convert to a flat data table}
\usage{
data2table(lr)
}
\arguments{
\item{lr}{the output of \code{calibrate} or \code{mergecal}}
}
\value{
a table that can be read by \code{IsoplotR}
}
\description{
trims the data to an \code{IsoplotR} \code{UPb}
    \code{type=5} table
}
\details{
converts an \code{IsoplotR} \code{type=9} object to a data
    table that is formatted as a \code{type=5} input table
}
\examples{
data(Cameca,package="simplex")
stand <- standards(dat=Cameca,prefix='Plesovice',tst=c(337.13,0.18))
calU <- calibration(stand=stand,oxide='UO2',parent='U238',
                    daughter='Pb206',cD4=18.7)
samp <- unknowns(dat=Cameca,prefix='Qinghu')
calUsamp <- calibrate(dat=samp,fit=calU)
calPbsamp <- getPbLogRatios(samp)
calsamp <- mergecal(calUsamp,calPbsamp)
tab <- data2table(calsamp)
}
