% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{data2table}
\alias{data2table}
\title{convert to a flat data table}
\usage{
data2table(x)
}
\arguments{
\item{x}{the output of \code{calibrate}}
}
\value{
a table n \code{IsoplotR} object of class \code{UPb}
    (\code{format=5})
}
\description{
trims the data to an \code{IsoplotR} \code{UPb}
    \code{type=5} table
}
\details{
converts an \code{IsoplotR} \code{type=9} object to a data
    table that is formatted as a \code{type=5} input table
}
\examples{
data(Cameca,package="simplex")
stand <- subset_samples(dat=Cameca,prefix='Plesovice')
fit <- calibration(stand)
samp <- subset_samples(dat=Cameca,prefix='Qinghu')
cal <- calibrate(samp,fit)
tab <- data2table(cal)
write.csv(tab,file='Qinghu.csv')
}
