% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{delta}
\alias{delta}
\title{calculate delta values}
\usage{
delta(cd, log = TRUE)
}
\arguments{
\item{cd}{an object of class \code{calibrated}}

\item{log}{use traditional delta notaion (\code{log=FALSE}) or the
logratio definition (\code{log=TRUE})?}
}
\value{
an object of class \code{delta}
}
\description{
convert stable isotope logratio data to delta notation
}
\examples{
data('oxygen',package='simplex')
dc <- drift(x=oxygen)
lr <- logratios(x=dc)
cal <- calibration(lr=lr,stand=standard(preset='NBS28'))
cd <- calibrate(cal)
del <- delta(cd)
tab <- data2table(del)
}
