% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{drift}
\alias{drift}
\title{drift correction}
\usage{
drift(x)
}
\arguments{
\item{x}{an object of class \code{simplex}}
}
\value{
an object of class \code{drift}
}
\description{
fits a generalised linear model to time resolved SIMS data
}
\examples{
data('SHRIMP_UPb',package='simplex')
dc <- drift(x=SHRIMP_UPb)
plot(dc,i=1)
}
