% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratios.R
\name{logratios}
\alias{logratios}
\title{logratio calculation}
\usage{
logratios(x, i = NULL)
}
\arguments{
\item{x}{an object of class \code{drift}}

\item{i}{a vector of mass spectrometer cycles to retain for the
logratio calculation. If \code{NULL}, uses the entire signal.}
}
\value{
an object of class \code{logratios}
}
\description{
extracts logratio averages from time resolved SIMS
    data
}
\examples{
\dontrun{
data('SHRIMP',package='simplex')
dc <- drift(x=SHRIMP)
lr <- logratios(dc)
plot(lr,i=1)
}
}
