% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{make_template}
\alias{make_template}
\title{Creates templates}
\usage{
make_template(fun = NULL, template_name = "basic")
}
\arguments{
\item{fun}{the function for which you want to make a template as a string.}

\item{template_name}{the resulting filename for your template as a string.}
}
\value{
A .r file in your working directory.
}
\description{
Creates templates
}
\examples{
make_template("boxplot", "triangle") # make any desired changes to the boxplot
a <- tibble(value = c(rnorm(2e2,1), 4*rbeta(2e2,1,4)), class = c(rep("a",2e2),rep("b",2e2)))
boxplot_template(x = a, template = "triangle.R")

}
