% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot2.R
\name{make_template}
\alias{make_template}
\title{Creates templates}
\usage{
make_template(fun = NULL, template_name = "basic")
}
\arguments{
\item{fun}{the function for which you want to make a template}

\item{template_name}{the resulting filename for your template.}
}
\value{
A .r file in your working directory.
}
\description{
Creates templates
}
\examples{
make_template("boxplot2", "triangle") # make any desired changes to the boxplot
a <- rbeta(1e2,1,3)
boxplot2_template(x = a, template = "triangle.R")

}
