% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{plot.calibrated}
\alias{plot.calibrated}
\title{plot calibrated data}
\usage{
\method{plot}{calibrated}(x, show.numbers = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{calibrated}}

\item{show.numbers}{logical. If \code{TRUE}, numbers the error
ellipses by aliquot.}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function}
}
\description{
shows the calibrated data on a logratio plot.
}
\examples{
\dontrun{
data('Cameca_UPb',package='simplex')
dc <- drift(x=Cameca_UPb)
lr <- logratios(x=dc)
st <- standard(preset='Temora-t')
p <- pairing(lr,stand=st)
cal <- calibration(lr=lr,pairing=p,stand=st,prefix='Tem')
cd <- calibrate(cal)
plot(cd)
}
}
