% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{plot.drift}
\alias{plot.drift}
\title{plot drift corrected data}
\usage{
\method{plot}{drift}(x, sname, i = 1, ...)
}
\arguments{
\item{x}{an object of class \code{drift}}

\item{i}{the sample number to be shown}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function.}

\item{snames}{the sample names to be shown}
}
\description{
shows the time resolved mass spectrometer signals
    fitted by a generalised linear model
}
\examples{
data('Cameca',package='simplex')
dc <- drift(x=Cameca)
plot(dc,i=1)
}
