% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratios.R
\name{plot.logratios}
\alias{plot.logratios}
\title{plot logratios}
\usage{
\method{plot}{logratios}(x, sname = NULL, i = 1, option = 1, ...)
}
\arguments{
\item{x}{an object of class \code{logratios}}

\item{sname}{the sample name to be shown}

\item{i}{the sample number to be shown}

\item{option}{if 1, plots the logratios against time, if 2, plots
the raw signals versus time. Both plots show the fitted values
as a solid line.}

\item{...}{optional arguments to be passed on to the generic
\code{plot} function.}
}
\description{
plot time resolved logratio data
}
\examples{
\dontrun{
data('SHRIMP',package='simplex')
dc <- drift(x=SHRIMP)
lr <- logratios(dc)
plot(lr,i=1,option=2)
}
}
