% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeresolved}
\alias{plot_timeresolved}
\title{plot the raw mass spectrometer data}
\usage{
plot_timeresolved(spot, ...)
}
\arguments{
\item{spot}{one item of a \code{simplex} list object}

\item{...}{optional parameters to the generic \code{plot} function}
}
\value{
a multi-panel plot of signal versus time
}
\description{
Shows the raw data for a single spot in a SIMS
    dataset.
}
\examples{
data(Cameca,package="simplex")
samp <- unknowns(dat=Cameca,prefix='Qinghu')
plot_timeresolved(spot=samp[[1]])
}
