% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeresolved}
\alias{plot_timeresolved}
\title{plot the raw mass spectrometer data}
\usage{
plot_timeresolved(samp, fit = NULL, ...)
}
\arguments{
\item{samp}{one item of a \code{simplex} list object}

\item{fit}{the output of \code{calibration}}

\item{...}{optional parameters to the generic \code{plot} function}
}
\value{
a multi-panel plot
}
\description{
Shows the raw data for a single spot in a SIMS
    dataset.
}
\examples{
data(Cameca,package="simplex")
stand <- standard(dat=Cameca,prefix='Plesovice',tst=c(337.13,0.18))
samp <- unknown(dat=Cameca,prefix='Qinghu')
cal <- calibration(stand,oxide='UO2')
plot_timeresolved(samp=samp[[1]],fit=cal)
}
