% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{SHRIMP}
\alias{SHRIMP}
\title{example SHRIMP dataset}
\description{
built-in Cameca dataset containing Temora and 91500
    zircon
}
\examples{
data(SHRIMP,package="simplex")
plot_timeresolved(SHRIMP[[1]])
}
