% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/documentation.R
\docType{package}
\name{simplex}
\alias{simplex}
\alias{simplex-package}
\title{Graphical User Interface for simplex}
\usage{
simplex(port = NULL)
}
\arguments{
\item{port}{The port on which to listen. If not provided, a random
unused port will be chosen and a browser window opened on that
port.}
}
\description{
Shinylight app for simplex

\code{simplex} is a data reduction package for Secondary Ion Mass
Spectrometry (SIMS), with functionality for both stable isotope
geochemistry and U-(Th)-Pb geochronology. The package accommodates
data from both Cameca and SHRIMP instruments. All mathematical
operations take place in logratio space (aka the simplex) to ensure
positivity and keep track of correlated uncertainties.
}
\details{
A list of documented functions may be viewed by typing
\code{help(package='simplex')}.  Source code and further
instructions are provided at
\url{https://github.com/pvermees/simplex/}.
}
\examples{
\donttest{simplex()}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pvermees/simplex}
}

}
\author{
\strong{Maintainer}: Pieter Vermeesch \email{p.vermeesch@ucl.ac.uk}

}
