% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{simplex2IsoplotR}
\alias{simplex2IsoplotR}
\title{convert to IsoplotR}
\usage{
simplex2IsoplotR(dat, format = 5)
}
\arguments{
\item{dat}{an object of class \code{calibrated}}

\item{format}{optional. Either 5 or 7, sets the format of the
\code{IsoplotR} oboject}
}
\value{
an object of class \code{UPb}
}
\description{
convert U-Pb or U-Th-Pb data to an IsoplotR object
}
\examples{
\dontrun{
m <- method('GA-UPb')
s <- standard(preset="Temora",prefix='TEM')
fname <- system.file('SHRIMP.pd',package='simplex')
cd <- process(f=fname,m=m,stand=s)
UPb <- simplex2IsoplotR(cd)
IsoplotR::concordia(UPb)
}
}
