% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{standard}
\alias{standard}
\title{define an isotopic reference standards}
\usage{
standard(preset, tst, measured, del, ref)
}
\arguments{
\item{preset}{(optional) text string. One of either

U-Pb age standards: \code{'Plesovice-t'}, \code{Qinghu-t},
\code{Penglai-t}, \code{'44069-t'}, \code{'Temora-t'},

O-isotope standards: \code{'Plesovice-O'}, \code{'NBS28-O'},
\code{'Qinghu-O'}, \code{'Penglai-O'},

or S-isotope standards: \code{'Sonora-S'}.}

\item{tst}{two element with the standard's age and its standard
error.}

\item{measured}{logical. Only relevant for geochronological data.
If \code{TRUE}, stores the composition of the standard as a
measured isotopic composition. If \code{FALSE}, stores the
composition as a mixture of common Pb and radiogenic Pb.}

\item{del}{Stable isotope composition of a standard in conventional
delta notation, specified as a two-item list containing a
vector \code{val} of \eqn{\delta}-values and its covariance
matrix \code{cov}.}

\item{ref}{Logratio composition of a reference standard such as
Vienna Standard Mean Ocean Water (VSMOW) or troilite. A
two-item list containing a vector \code{val} of logratio-values
and its covariance matrix \code{cov}.}
}
\value{
an object of class \code{standard}
}
\description{
specify the isotopic composition of a reference
    material for SIMS data calibration.
}
\examples{
data(Cameca_UPb,package="simplex")
st <- standard(preset="Plesovice-t")
cal <- calibration(lr=Cameca_UPb,stand=st)
plot(cal)
}
