% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{standard}
\alias{standard}
\title{define an isotopic reference standards}
\usage{
standard(
  preset,
  prefix = preset,
  tst,
  val,
  cov = matrix(0, length(val), length(val)),
  common
)
}
\arguments{
\item{preset}{(optional) text string. One of either
\code{'Plesovice'}, \code{'44069'}, \code{'Temora'}, or
\code{'NBS28'}.}

\item{prefix}{text string marking the first characters of the
sample names that mark the standard analyses}

\item{tst}{(optional) two-element vector with the age and standard
error of the (presumed concordant) age standard and its
analytical uncertainty.}

\item{val}{(optional) true isotopic composition of
\eqn{^{206}}Pb/\eqn{^{238}}U- and
\eqn{^{208}}Pb/\eqn{^{232}}Th-ratios of the U-Pb age standard,
or true \eqn{\delta^{18}O} and \eqn{\delta^{16}O} values of the
oxygen isotope reference material.}

\item{cov}{(optional) the covariance matrix of \code{val}}

\item{common}{(optional) the common-Pb composition
(\eqn{^{206}}Pb/\eqn{^{204}}Pb and
\eqn{^{208}}Pb/\eqn{^{204}}Pb).}
}
\value{
an object of class \code{standard}
}
\description{
specify the isotopic composition of a reference
    material for SIMS data calibration.
}
\examples{
data(Cameca,package="simplex")
dc <- drift(x=Cameca)
lr <- logratios(x=dc)
st <- standard(preset="Plesovice")
cal <- calibration(lr=lr,stand=st)
plot(cal,option=3)
}
