% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{standards}
\alias{standards}
\title{define the standards in a dataset}
\usage{
standards(dat, prefix, invert = FALSE)
}
\arguments{
\item{dat}{an object of class \code{simplex}}

\item{prefix}{text string to match}

\item{invert}{logical.  If \code{TRUE} return samples whose names
do _not_ match}
}
\value{
an object of class \code{standards}
}
\description{
select a subset of standards from a \code{simplex}
    dataset.
}
\examples{
data(Cameca,package="simplex")
stand <- standards(Cameca,prefix='Plesovice')
}
