% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{subset_samples}
\alias{subset_samples}
\title{subset a dataset of class \code{simplex}}
\usage{
subset_samples(dat, prefix = "Plesovice")
}
\arguments{
\item{dat}{an object of class \code{simplex}}

\item{prefix}{text string to match}
}
\value{
an object of class \code{simplex}
}
\description{
select a subset of samples or standards from a
    \code{simplex} dataset.
}
\examples{
data(Cameca,package="simplex")
stand <- subset_samples(Cameca,prefix='Plesovice')
}
